+++ NEWS.md
@@ -1,6 +1,7 @@
 # httr2 (development version)
 
 * `req_perform_promise()` upgraded to use event-driven async based on waiting efficiently on curl socket activity (#579).
+* New `req_oauth_token_exchange()` and `oauth_flow_token_exchange()` functions implement the OAuth token exchange protocol from RFC 8693 (@atheriel, #460).
 
 # httr2 1.0.6